pyproject.toml
[project]
name = "redsoft-test-task"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django-split-settings>=1.3.2",
    "django>=5.1",
    "djangorestframework-simplejwt>=5.3.1",
    "djangorestframework>=3.15.2",
    "drf-spectacular>=0.27.2",
    "psycopg2-binary>=2.9.9",
    "pycryptodome>=3.20.0",
    "ruff>=0.6.3",
]
